+++ src/main/java/com/javajober/template/dto/MemberAuthResponse.java
@@ -4,8 +4,8 @@ package com.javajober.template.dto;
 import java.util.List;
 
 import com.javajober.member.domain.Member;
-import com.javajober.entity.MemberGroup;
-import com.javajober.entity.TemplateAuth;
+import com.javajober.template.domain.MemberGroup;
+import com.javajober.template.domain.TemplateAuth;
 
 import lombok.Builder;
 import lombok.Getter;
