+++ src/main/java/com/javajober/template/repository/TemplateAuthRepository.java
@@ -5,8 +5,8 @@ import java.util.Optional;
 
 import com.javajober.core.error.exception.Exception404;
 import com.javajober.core.message.ErrorMessage;
-import com.javajober.entity.TemplateAuth;
-import com.javajober.entity.TemplateBlock;
+import com.javajober.template.domain.TemplateAuth;
+import com.javajober.template.domain.TemplateBlock;
 
 import org.springframework.data.repository.Repository;
