+++ src/main/java/com/javajober/template/service/TemplateBlockService.java
@@ -8,9 +8,9 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.javajober.core.error.exception.Exception404;
 import com.javajober.core.message.ErrorMessage;
-import com.javajober.entity.MemberGroup;
-import com.javajober.entity.TemplateAuth;
-import com.javajober.entity.TemplateBlock;
+import com.javajober.template.domain.MemberGroup;
+import com.javajober.template.domain.TemplateAuth;
+import com.javajober.template.domain.TemplateBlock;
 import com.javajober.template.dto.TemplateBlockRequest;
 import com.javajober.template.dto.TemplateBlockRequests;
 import com.javajober.template.dto.TemplateBlockResponse;
