+++ src/main/java/com/javajober/template/service/TemplateService.java
@@ -11,14 +11,14 @@ import com.javajober.core.error.exception.Exception404;
 import com.javajober.core.message.ErrorMessage;
 import com.javajober.entity.SpaceWallCategory;
 import com.javajober.entity.SpaceWallCategoryType;
-import com.javajober.entity.Template;
-import com.javajober.entity.TemplateBlock;
+import com.javajober.template.domain.Template;
+import com.javajober.template.domain.TemplateBlock;
 import com.javajober.template.dto.MemberAuthResponse;
 import com.javajober.entity.AddSpace;
 import com.javajober.entity.Member;
-import com.javajober.entity.MemberGroup;
+import com.javajober.template.domain.MemberGroup;
 import com.javajober.entity.SpaceType;
-import com.javajober.entity.TemplateAuth;
+import com.javajober.template.domain.TemplateAuth;
 import com.javajober.template.dto.TemplateResponse;
 import com.javajober.template.repository.MemberGroupRepository;
 import com.javajober.template.repository.SpaceWallCategoryRepository;
